fix: move sourceselector to bbgo folder
